requirements.txt
langchain
langchain-openai
langchain-community
langgraph
faiss-cpu
rank_bm25
sentence-transformers
fastapi
uvicorn
pydantic
numpy
python-dotenv
tiktoken